pyproject.toml
[project]
name = "sbaas-productivity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.7",
    "SQLAlchemy>=2.0",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=24.0",
    "mypy>=1.8",
    "pytest-cov>=4.1",
    "pylint>=3.0",
]
